Separate the compact proof store from proofs
CompactProofStore is serializable (e.g., writable to disk in compact form)
and ProofStore is (hypothetically) usable for constructing proofs.

The tests show some basic usage examples. They prove that A) the root
hash can be reconstructed (should be obvious from the tree construction)
and B) the CompactProofStore is guaranteed to be smaller than a naive
list of proofs.

Given A, all proofs can be created and verified. Though the code for it
hasn't been written yet.